+++ src/constants/defaultProps.ts
@@ -1,18 +1,20 @@
+import { type SVGDefaultProps } from '../interfaces/misc'
+
 /**
  * Default properties for SVG components.
  * This object provides default values for various properties of SVG components.
  */
-export const defaultProps: Record<string, any> = {
-  size: 28,
-  width: 28,
-  height: 28,
-  fill: 'rgb(0, 0, 0)',
+export const defaultProps: SVGDefaultProps = {
+  className: '',
   color: 'rgb(0, 0, 0)',
+  fill: 'rgb(0, 0, 0)',
+  height: 28,
+  size: 28,
+  state: false,
   stroke: 'rgb(0, 0, 0)',
   strokeColor: 'rgb(0, 0, 0)',
   strokeWidth: 0.2,
-  state: false,
   style: {},
   sx: {},
-  className: '',
+  width: 28,
 }