+++ src/interfaces/svgExports.d.ts
@@ -51,6 +51,7 @@ export interface SvgComponent {
   /**
    * Parameters related to the SVG component, which are passed as props to the component.
    */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   params: Record<string, any>
 }
 
@@ -69,6 +70,7 @@ export interface SvgComponentDetails {
   /**
    * An object containing the props of the component.
    */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   props: Record<string, any>
   /**
    * A flag indicating whether the component uses framer-motion