+++ src/utilities/fileSystem/saveResponseFiles.ts
@@ -1,14 +1,17 @@
-import { Uri, commands, window, workspace } from 'vscode'
 import * as fs from 'fs'
 
-export async function saveResponseFile(data: object) {
+import { Uri, commands, window, workspace } from 'vscode'
+
+import { isEmpty } from '../misc'
+
+export async function saveResponseFile(data: object, path?: string): Promise<void> {
   const content = JSON.stringify(data, null, 2)
-  const filePath = workspace.rootPath + '/responseFile.json'
+  const filePath = path ?? workspace.workspaceFolders?.[0].uri.fsPath + '/response.json'
 
   try {
     await new Promise((resolve, reject) => {
       fs.writeFile(filePath, content, (err) => {
-        if (err) {
+        if (!isEmpty(err)) {
           reject(err)
         } else {
           resolve('File saved successfully.')