+++ src/utilities/svg/extractSVGComponentExports.ts
@@ -1,9 +1,11 @@
-/* eslint-disable @typescript-eslint/naming-convention */
 import traverse from '@babel/traverse'
 import * as t from '@babel/types'
 
+import { type ANY_TYPE } from '../../interfaces/misc'
 import { type SvgComponent, type SvgExport } from '../../interfaces/svgExports'
 import { parseFileContent } from '../babelParser'
+import { isEmpty } from '../misc'
+
 import { extractSvgComponentFromNode } from './exportParser'
 
 /**
@@ -13,12 +15,13 @@ import { extractSvgComponentFromNode } from './exportParser'
  */
 export async function extractSVGComponentExports(filePath: string): Promise<{
   result: SvgExport
-  base: Record<string, { declaration: any; params: Record<string, any> }>
+  base: Record<string, { declaration: ANY_TYPE; params: Record<string, ANY_TYPE> }>
 }> {
   try {
     // Parse the file content into an AST (Abstract Syntax Tree)
     const ast = parseFileContent(filePath)
-    const baseExports: Record<string, { declaration: any; params: Record<string, any> }> = {}
+    const baseExports: Record<string, { declaration: ANY_TYPE; params: Record<string, ANY_TYPE> }> =
+      {}
     const exports: SvgComponent[] = []
     const notExports: SvgComponent[] = []
     const identifiers = new Set<string>()
@@ -29,13 +32,13 @@ export async function extractSVGComponentExports(filePath: string): Promise<{
       Declaration(path) {
         const { node, parent } = path
 
-        if (t.isProgram(parent) && node) {
-          let declaration: any = node
+        if (t.isProgram(parent) && !isEmpty(node)) {
+          let declaration: ANY_TYPE = node
           let isExported = false
 
           if (
             (t.isExportNamedDeclaration(node) || t.isExportDefaultDeclaration(node)) &&
-            node.declaration
+            !isEmpty(node.declaration)
           ) {
             declaration = node.declaration
             isExported = true
@@ -44,7 +47,7 @@ export async function extractSVGComponentExports(filePath: string): Promise<{
           if (t.isFunctionDeclaration(declaration)) {
             extractSvgComponentFromNode(declaration as t.Declaration, 'function')
               .then((result) => {
-                if (result) {
+                if (result !== undefined) {
                   // Exported function declaration 'export function functionName() {}'
                   if (isExported || identifiers.has(result.name)) {
                     lengthExports++
@@ -64,7 +67,7 @@ export async function extractSVGComponentExports(filePath: string): Promise<{
               if (t.isIdentifier(d.id)) {
                 extractSvgComponentFromNode(d, 'variable')
                   .then((result) => {
-                    if (result) {
+                    if (result !== undefined) {
                       // Exported variable declaration 'export const variableName = value;'
                       if (isExported || identifiers.has(result.name)) {
                         lengthExports++
@@ -89,7 +92,7 @@ export async function extractSVGComponentExports(filePath: string): Promise<{
                 identifiers.add(property.key.name)
               }
             })
-          } else if (t.isExportNamedDeclaration(declaration) && declaration.specifiers) {
+          } else if (t.isExportNamedDeclaration(declaration) && !isEmpty(declaration.specifiers)) {
             declaration.specifiers.forEach((specifier) => {
               if (t.isExportSpecifier(specifier) && t.isIdentifier(specifier.exported)) {
                 identifiers.add(specifier.exported.name)