+++ src/utilities/svg/propertyManager.ts
@@ -1,15 +1,16 @@
 import { REST_PROPS_KEY } from '../../constants/misc'
+import { type ANY_TYPE } from '../../interfaces/misc'
 
 class PropertyManager {
-  private properties: Record<string, any> = {}
+  private properties: Record<string, ANY_TYPE> = {}
 
   /**
    * Gets a specific property or all properties if no key is specified.
    * @param {string} key - Key of the property to retrieve (optional).
    * @returns {any} - The value of the property or an object with all properties.
    */
-  get(key?: string): any {
-    if (key) {
+  get(key?: string): ANY_TYPE {
+    if (key !== undefined) {
       return this.properties[key]
     } else {
       return this.properties
@@ -20,7 +21,7 @@ class PropertyManager {
    * Retrieves the rest properties from the property manager.
    * @returns A record containing the rest properties.
    */
-  getRestProps(): Record<string, any> {
+  getRestProps(): Record<string, ANY_TYPE> {
     return this.properties[REST_PROPS_KEY]
   }
 
@@ -29,7 +30,7 @@ class PropertyManager {
    * @param {string} key - Key of the property.
    * @param {any} value - Value of the property.
    */
-  set(key: string, value: any): void {
+  set(key: string, value: ANY_TYPE): void {
     this.properties[key] = value
   }
 
@@ -37,7 +38,7 @@ class PropertyManager {
    * Sets the value of multiple properties.
    * @param {Record<string, any>} properties - Object with the properties to set.
    */
-  setAll(properties: Record<string, any>): void {
+  setAll(properties: Record<string, ANY_TYPE>): void {
     this.properties = properties
   }
 
@@ -46,7 +47,7 @@ class PropertyManager {
    * @param {string} key - Key of the property to clear (optional).
    */
   clean(key?: string): void {
-    if (key) {
+    if (key !== undefined) {
       if (key in this.properties) {
         const { [key]: omitted, ...rest } = this.properties
         this.properties = rest