+++ src/utilities/vscode/theme.ts
@@ -1,6 +1,7 @@
+import * as path from 'path'
+
 import * as Color from 'color'
 import { extensions, window } from 'vscode'
-import * as path from 'path'
 
 import {
   type IStandaloneThemeData,
@@ -21,7 +22,7 @@ export function getCurrentTheme(): ThemeMode {
   // Get the current color theme's kind value
   const currentTheme = window.activeColorTheme.kind
 
-  /// Check if the kind value corresponds to a "light" theme (1, 4) or "dark" theme (2, 3)
+  // Check if the kind value corresponds to a "light" theme (1, 4) or "dark" theme (2, 3)
   // and return the appropriate theme name
   return currentTheme === 1 || currentTheme === 4 ? 'light' : 'dark'
 }
@@ -54,9 +55,9 @@ function themeParser(theme: string, mode: ThemeMode, values: IThemeObject): IThe
         token: scope.trim(),
       }
 
-      if (foreground) rule.foreground = (Color(foreground).hex() ?? '').toLowerCase()
-      if (background) rule.background = (Color(background).hex() ?? '').toLowerCase()
-      if (fontStyle) rule.fontStyle = fontStyle
+      if (foreground !== undefined) rule.foreground = (Color(foreground).hex() ?? '').toLowerCase()
+      if (background !== undefined) rule.background = (Color(background).hex() ?? '').toLowerCase()
+      if (fontStyle !== undefined) rule.fontStyle = fontStyle
 
       rules.push(rule)
     }
@@ -80,11 +81,11 @@ function themeParser(theme: string, mode: ThemeMode, values: IThemeObject): IThe
  * @param colorTheme The name or ID of the color theme.
  * @returns A promise that resolves to the theme data, or undefined if the theme is not found.
  */
-export async function getThemeInfo(colorTheme: string) {
+export async function getThemeInfo(colorTheme: string): Promise<IThemeData | undefined> {
   // Check if the theme is cached
   const cachedTheme = themeCache.get(colorTheme)
 
-  if (cachedTheme) {
+  if (cachedTheme !== undefined) {
     return cachedTheme
   }
 
@@ -100,7 +101,7 @@ export async function getThemeInfo(colorTheme: string) {
     )
   })
 
-  if (themeExtensionProvided) {
+  if (themeExtensionProvided !== undefined) {
     const { extensionPath, packageJSON } = themeExtensionProvided
 
     // Theme provided by an extension
@@ -108,7 +109,7 @@ export async function getThemeInfo(colorTheme: string) {
       (theme: ThemeInfo) => theme.label === colorTheme || theme.id === colorTheme
     )
 
-    if (themeInfo?.path) {
+    if (themeInfo?.path !== undefined) {
       const themePath = path.join(extensionPath, themeInfo.path)
       const themeJSON = await import(themePath)
 
@@ -116,7 +117,7 @@ export async function getThemeInfo(colorTheme: string) {
     }
   }
 
-  if (result) {
+  if (result !== undefined) {
     themeCache.set(colorTheme, result)
   }